updating for multiple OS

+++ stack/files/standard.sh
@@ -1,6 +1,14 @@
 #!/bin/bash
-sudo apt update -yq
-sudo DEBIAN_FRONTEND=noninteractive apt upgrade -yq
-sudo apt install wget curl git python3-minimal -yq
-#fix for python3 symlink
-#ln -s /usr/bin/python3.5 /usr/bin/python
+
+if [ -f /etc/redhat-release ]; then
+  echo "######################### UPDATES #############################"
+  sudo yum update -y
+  echo "######################### INSTALLS #############################"
+  sudo yum install wget curl git python3 -y
+else
+  echo "######################### UPDATES #############################"
+  sudo apt update -yq
+  echo "######################### INSTALLS #############################"
+  sudo DEBIAN_FRONTEND=noninteractive apt upgrade -yq
+  sudo apt install wget curl git python3-minimal -yq
+fi
